# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FileBrowser Backend API",
    "description": "Full API reference and interactive testing interface for the FileBrowser Node.js/Express backend.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base URL"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "XAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the JWT token (without 'Bearer ' prefix)."
      },
      "XAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Auth",
        "description": "Alternative authorization header using 'X-Auth'."
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "scope": { "type": "string" },
          "locale": { "type": "string" },
          "viewMode": { "type": "string" },
          "singleClick": { "type": "boolean" },
          "lockPassword": { "type": "boolean" },
          "hideDotfiles": { "type": "boolean" },
          "dateFormat": { "type": "boolean" },
          "commands": {
            "type": "array",
            "items": { "type": "string" }
          },
          "rules": {
            "type": "array",
            "items": { "type": "string" }
          },
          "perm": {
            "type": "object",
            "properties": {
              "admin": { "type": "boolean" },
              "execute": { "type": "boolean" },
              "create": { "type": "boolean" },
              "rename": { "type": "boolean" },
              "modify": { "type": "boolean" },
              "delete": { "type": "boolean" },
              "share": { "type": "boolean" },
              "download": { "type": "boolean" }
            }
          },
          "createdAt": { "type": "integer" },
          "updatedAt": { "type": "integer" }
        }
      },
...
    }
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Admin: List all users",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Admin: Create a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "scope": { "type": "string" },
                  "locale": { "type": "string" },
                  "viewMode": { "type": "string", "enum": ["mosaic", "list"] },
                  "commands": { "type": "array", "items": { "type": "string" } },
                  "perm": {
                    "type": "object",
                    "properties": {
                      "admin": { "type": "boolean" },
                      "execute": { "type": "boolean" },
                      "create": { "type": "boolean" },
                      "rename": { "type": "boolean" },
                      "modify": { "type": "boolean" },
                      "delete": { "type": "boolean" },
                      "share": { "type": "boolean" },
                      "download": { "type": "boolean" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update user details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "currentPassword": { "type": "string" },
                  "scope": { "type": "string" },
                  "locale": { "type": "string" },
                  "viewMode": { "type": "string" },
                  "singleClick": { "type": "boolean" },
                  "hideDotfiles": { "type": "boolean" },
                  "dateFormat": { "type": "boolean" },
                  "lockPassword": { "type": "boolean" },
                  "commands": { "type": "array", "items": { "type": "string" } },
                  "rules": { "type": "array", "items": { "type": "string" } },
                  "perm": {
                    "type": "object",
                    "properties": {
                      "admin": { "type": "boolean" },
                      "execute": { "type": "boolean" },
                      "create": { "type": "boolean" },
                      "rename": { "type": "boolean" },
                      "modify": { "type": "boolean" },
                      "delete": { "type": "boolean" },
                      "share": { "type": "boolean" },
                      "download": { "type": "boolean" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentPassword": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Deleted successfully" }
        }
      }
    },
...
  }
}